update documentations of reflection, bridge and jsbBridgeWrapper (#12097)
